// HelpDeskFE/HelpDeskFETp/src/api/model/createDeviceDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * HelpDesk API
 * OpenAPI spec version: v1
 */

export interface CreateDeviceDto {
  quantity?: number;
  /** @nullable */
  brandName?: string | null;
  /** @nullable */
  code?: string | null;
  idDeviceType?: number;
  idUser?: number;
  idArea?: number;
  /** @nullable */
  observation?: string | null;
  isActive?: boolean;
}