// HelpDeskFE/HelpDeskFETp/src/api/model/updateDeviceDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * HelpDesk API
 * OpenAPI spec version: v1
 */

export interface UpdateDeviceDto {
  quantity?: number;
  /** @nullable */
  brandName?: string | null;
  /** @nullable */
  code?: string | null;
  idDeviceType?: number;
  idUser?: number;
  idArea?: number;
  /** @nullable */
  observation?: string | null;
  isActive?: boolean;
}